Update for mac

+++ CxImage/ximajas.h
@@ -64,9 +64,9 @@ protected:
 			// <vho> - end
 		}
 		
-		static int JasRead(jas_stream_obj_t *obj, char *buf, unsigned cnt)
+		static long JasRead(jas_stream_obj_t *obj, char *buf, size_t cnt)
 		{		return ((CxFile*)obj)->Read(buf,1,cnt); }
-		static int JasWrite(jas_stream_obj_t *obj, const char *buf, unsigned cnt)
+		static long JasWrite(jas_stream_obj_t *obj, const char *buf, size_t cnt)
 		{		return ((CxFile*)obj)->Write(buf,1,cnt); }
 		static long JasSeek(jas_stream_obj_t *obj, long offset, int origin)
 		{		return ((CxFile*)obj)->Seek(offset,origin); }
